+++ fixtures/src/a-element.js
@@ -4,9 +4,17 @@
 
 console.log(`
 <md-icon-button>
-  <md-icon>test</md-icon>
+  <md-icon>soap</md-icon>
+</md-icon-button>
+<md-icon-button>
+  <md-icon>delete</md-icon>
+</md-icon-button>
+<md-icon-button>
+  <md-icon>fake_name</md-icon>
 </md-icon-button>
 
 <!-- the following element is in html comment, therefore shouldn't be included in the final builds -->
 <!-- <md-chip-set></md-chip-set> -->
 `);
+
+// <md-icon>add</md-icon>